Replace HXT with libxml
The primary issue is that str:encode-uri function from EXSLT is not
available with HXT (but available with libxml's friends: libxslt and
libexslt, which are generally more complete). This function is
important for HTML documents.

Another reason is libxml being better documented and somewhat easier
to work with.

Yet another reason to switch is the intent to avoid dependencies that
reimplement common functionality, and/or normally get statically
linked.